refactor(api): move reset password demand repository
