+++ api/src/identity-access-management/infrastructure/repositories/reset-password-demand.repository.js
@@ -1,5 +1,5 @@
-import { PasswordResetDemandNotFoundError } from '../../../lib/domain/errors.js';
-import { ResetPasswordDemand } from '../orm-models/ResetPasswordDemand.js';
+import { PasswordResetDemandNotFoundError } from '../../../../lib/domain/errors.js';
+import { ResetPasswordDemand } from '../../../../lib/infrastructure/orm-models/ResetPasswordDemand.js';
 
 const create = function (demand) {
   return new ResetPasswordDemand(demand).save();
